// src/app/enums/navigation.enum.ts
export const enum PermissionType {
    Administration = 1, 
    UserList = 2,
    Profile = 3,
    Annulment = 4,
    History = 5,
    Discount = 6,
    DiscountList = 7,

    Offices = 10,
    Office = 11,
    Movements = 12,
    Sales = 13,
    Cancel = 14,
    Changes = 15,
    OfficeList = 16,
    SaleList = 17,

    Stock = 20,
    Items = 21,
    Dispatch = 22,
    DispatchList = 23
   
  }